selectors

Module bloom

Counting and non-counting Bloom filters tuned for use as ancestor filters for selector matching.

Structs

BloomStorageBool

Storage class for a CountingBloomFilter that has 1-bit counters.

BloomStorageU8

Storage class for a CountingBloomFilter that has 8-bit counters.

CountingBloomFilter

A counting Bloom filter with parameterized storage to handle counters of different sizes. For now we assume that having two hash functions is enough, but we may revisit that decision later.

Constants

ARRAY_SIZE 🔒

BLOOM_HASH_MASK

KEY_MASK 🔒

KEY_SIZE 🔒

Traits

BloomStorage

Functions

hash1 🔒

hash2 🔒

Type Aliases

BloomFilter

A counting Bloom filter with 8-bit counters.