Type Alias selectors::bloom::BloomFilter

pub type BloomFilter = CountingBloomFilter<BloomStorageU8>;

A counting Bloom filter with 8-bit counters.

Aliased Type

struct BloomFilter { /* private fields */ }

Implementations

impl<S> CountingBloomFilter<S>where S: BloomStorage,

pub fn new() -> Self

Creates a new bloom filter.

pub fn clear(&mut self)

pub fn is_zeroed(&self) -> bool

pub fn insert_hash(&mut self, hash: u32)

Inserts an item with a particular hash into the bloom filter.

pub fn remove_hash(&mut self, hash: u32)

Removes an item with a particular hash from the bloom filter.

pub fn might_contain_hash(&self, hash: u32) -> bool

Check whether the filter might contain an item with the given hash. This can sometimes return true even if the item is not in the filter, but will never return false for items that are actually in the filter.