selectors

Module context

Structs

MatchingContext

Data associated with the matching process for a element. This context is used across many selectors for an element, so it’s not appropriate for transient data that applies to only a single selector.

SelectorCaches

Set of caches (And cache-likes) that speed up expensive selector matches.

Enums

MatchingForInvalidation

Whether we’re matching in the contect of invalidation.

MatchingMode

What kind of selector matching mode we should use.

NeedsSelectorFlags

Whether we need to set selector invalidation flags on elements for this match request.

QuirksMode

Which quirks mode is this document in.

VisitedHandlingMode

The mode to use when matching unvisited and visited links.