Enum selectors::context::VisitedHandlingMode

pub enum VisitedHandlingMode {
    AllLinksUnvisited,
    AllLinksVisitedAndUnvisited,
    RelevantLinkVisited,
}

The mode to use when matching unvisited and visited links.

Variants
AllLinksUnvisited

All links are matched as if they are unvisted.

AllLinksVisitedAndUnvisited

All links are matched as if they are visited and unvisited (both :link and :visited match).

This is intended to be used from invalidation code, to be conservative about whether we need to restyle a link.

RelevantLinkVisited

A element's "relevant link" is the element being matched if it is a link or the nearest ancestor link. The relevant link is matched as though it is visited, and all other links are matched as if they are unvisited.

Methods

impl VisitedHandlingMode

pub fn matches_visited(&self) -> bool

pub fn matches_unvisited(&self) -> bool