selectors::context

Struct MatchingContext

pub struct MatchingContext<'a, Impl>
where
    Impl: SelectorImpl,
{

    matching_mode: MatchingMode,
    pub bloom_filter: Option<&'a BloomFilter>,
    pub scope_element: Option<OpaqueElement>,
    pub current_host: Option<OpaqueElement>,
    visited_handling: VisitedHandlingMode,
    pub featureless: bool,
    nesting_level: usize,
    in_negation: bool,
    pub pseudo_element_matching_fn: Option<&'a dyn Fn(&Impl::PseudoElement) -> bool>,
    pub extra_data: Impl::ExtraMatchingData<'a>,
    current_relative_selector_anchor: Option<OpaqueElement>,
    quirks_mode: QuirksMode,
    needs_selector_flags: NeedsSelectorFlags,
    matching_for_invalidation: MatchingForInvalidation,
    matching_for_revalidation: bool,
    pub selector_caches: &'a mut SelectorCaches,
    classes_and_ids_case_sensitivity: CaseSensitivity,
    _impl: PhantomData<Impl>,

}

Data associated with the matching process for a element. This context is used across many selectors for an element, so it’s not appropriate for transient data that applies to only a single selector.

Fields
matching_mode: MatchingMode

Input with the matching mode we should use when matching selectors.

bloom_filter: Option<&'a BloomFilter>

Input with the bloom filter used to fast-reject selectors.

scope_element: Option<OpaqueElement>

The element which is going to match :scope pseudo-class. It can be either one :scope element, or the scoping element.

Note that, although in theory there can be multiple :scope elements, in current specs, at most one is specified, and when there is one, scoping element is not relevant anymore, so we use a single field for them.

When this is None, :scope will match the root element.

See https://drafts.csswg.org/selectors-4/#scope-pseudo

current_host: Option<OpaqueElement>

The current shadow host we’re collecting :host rules for.

visited_handling: VisitedHandlingMode

Controls how matching for links is handled.

featureless: bool

Whether we’re currently matching a featureless element.

nesting_level: usize

The current nesting level of selectors that we’re matching.

in_negation: bool

Whether we’re inside a negation or not.

pseudo_element_matching_fn: Option<&'a dyn Fn(&Impl::PseudoElement) -> bool>

An optional hook function for checking whether a pseudo-element should match when matching_mode is ForStatelessPseudoElement.

extra_data: Impl::ExtraMatchingData<'a>

Extra implementation-dependent matching data.

current_relative_selector_anchor: Option<OpaqueElement>

The current element we’re anchoring on for evaluating the relative selector.

quirks_mode: QuirksModeneeds_selector_flags: NeedsSelectorFlagsmatching_for_invalidation: MatchingForInvalidation

Whether we’re matching in the context of invalidation.

matching_for_revalidation: bool

Whether we’re matching in the context of revalidation.

selector_caches: &'a mut SelectorCaches

Caches to speed up expensive selector matches.

classes_and_ids_case_sensitivity: CaseSensitivity_impl: PhantomData<Impl>
Implementations

impl<'a, Impl> MatchingContext<'a, Impl>
where Impl: SelectorImpl,

pub fn new( matching_mode: MatchingMode, bloom_filter: Option<&'a BloomFilter>, selector_caches: &'a mut SelectorCaches, quirks_mode: QuirksMode, needs_selector_flags: NeedsSelectorFlags, matching_for_invalidation: MatchingForInvalidation, ) -> Self

Constructs a new MatchingContext.

pub fn new_for_revalidation( bloom_filter: Option<&'a BloomFilter>, selector_caches: &'a mut SelectorCaches, quirks_mode: QuirksMode, needs_selector_flags: NeedsSelectorFlags, ) -> Self

Constructs a new MatchingContext for revalidation.

pub fn new_for_visited( matching_mode: MatchingMode, bloom_filter: Option<&'a BloomFilter>, selector_caches: &'a mut SelectorCaches, visited_handling: VisitedHandlingMode, quirks_mode: QuirksMode, needs_selector_flags: NeedsSelectorFlags, matching_for_invalidation: MatchingForInvalidation, ) -> Self

Constructs a new MatchingContext for use in visited matching.

fn new_internal( matching_mode: MatchingMode, bloom_filter: Option<&'a BloomFilter>, selector_caches: &'a mut SelectorCaches, visited_handling: VisitedHandlingMode, quirks_mode: QuirksMode, needs_selector_flags: NeedsSelectorFlags, matching_for_invalidation: MatchingForInvalidation, matching_for_revalidation: bool, ) -> Self

pub fn nth_index_cache( &mut self, is_of_type: bool, is_from_end: bool, selectors: &[Selector<Impl>], ) -> &mut NthIndexCacheInner

pub fn is_nested(&self) -> bool

Whether we’re matching a nested selector.

pub fn in_negation(&self) -> bool

Whether we’re matching inside a :not(..) selector.

pub fn quirks_mode(&self) -> QuirksMode

The quirks mode of the document.

pub fn matching_mode(&self) -> MatchingMode

The matching-mode for this selector-matching operation.

pub fn needs_selector_flags(&self) -> bool

Whether we need to set selector flags.

pub fn matching_for_invalidation(&self) -> bool

Whether or not we’re matching to invalidate.

pub fn matching_for_revalidation(&self) -> bool

Whether or not we’re matching to revalidate.

pub fn matching_for_invalidation_comparison(&self) -> Option<bool>

Whether or not we’re comparing for invalidation, if we are matching for invalidation.

pub fn for_invalidation_comparison<F, R>(&mut self, f: F) -> R
where F: FnOnce(&mut Self) -> R,

Run the given matching function for before/after invalidation comparison.

pub fn classes_and_ids_case_sensitivity(&self) -> CaseSensitivity

The case-sensitivity for class and ID selectors

pub fn nest<F, R>(&mut self, f: F) -> R
where F: FnOnce(&mut Self) -> R,

Runs F with a deeper nesting level.

pub fn nest_for_negation<F, R>(&mut self, f: F) -> R
where F: FnOnce(&mut Self) -> R,

Runs F with a deeper nesting level, and marking ourselves in a negation, for a :not(..) selector, for example.

pub fn visited_handling(&self) -> VisitedHandlingMode

pub fn with_featureless<F, R>(&mut self, featureless: bool, f: F) -> R
where F: FnOnce(&mut Self) -> R,

Runs F with a different featureless element flag.

pub fn featureless(&self) -> bool

Returns whether the currently matching element is acting as a featureless element (e.g., because we’ve crossed a shadow boundary). This is used to implement the :host selector rules properly.

pub fn with_visited_handling_mode<F, R>( &mut self, handling_mode: VisitedHandlingMode, f: F, ) -> R
where F: FnOnce(&mut Self) -> R,

Runs F with a different VisitedHandlingMode.

pub fn with_shadow_host<F, E, R>(&mut self, host: Option<E>, f: F) -> R
where E: Element, F: FnOnce(&mut Self) -> R,

Runs F with a given shadow host which is the root of the tree whose rules we’re matching.

pub fn shadow_host(&self) -> Option<OpaqueElement>

Returns the current shadow host whose shadow root we’re matching rules against.

pub fn nest_for_relative_selector<F, R>( &mut self, anchor: OpaqueElement, f: F, ) -> R
where F: FnOnce(&mut Self) -> R,

Runs F with a deeper nesting level, with the given element as the anchor, for a :has(…) selector, for example.

pub fn nest_for_scope<F, R>(&mut self, scope: Option<OpaqueElement>, f: F) -> R
where F: FnOnce(&mut Self) -> R,

Runs F with a deeper nesting level, with the given element as the scope.

pub fn nest_for_scope_condition<F, R>( &mut self, scope: Option<OpaqueElement>, f: F, ) -> R
where F: FnOnce(&mut Self) -> R,

Runs F with a deeper nesting level, with the given element as the scope, for matching scope-start and/or scope-end conditions.

pub fn relative_selector_anchor(&self) -> Option<OpaqueElement>

Returns the current anchor element to evaluate the relative selector against.