Enum selectors::matching::CompoundSelectorMatchingResult

pub enum CompoundSelectorMatchingResult {
    FullyMatched,
    Matched {
        next_combinator_offset: usize,
    },
    NotMatched,
}

Whether a compound selector matched, and whether it was the rightmost selector inside the complex selector.

Variants
FullyMatched

The selector was fully matched.

Matched

The compound selector matched, and the next combinator offset is next_combinator_offset.

Fields of Matched

next_combinator_offset: usize

NotMatched

The selector didn't match.