Function selectors::matching::can_fast_reject

fn can_fast_reject<E>(selector: &CompoundSelector<E::Impl>, element: &E, parent_bf: Option<&BloomFilter>, shareable: &mut bool) -> Option<SelectorMatchingResult> where E: Element

Quickly figures out whether or not the compound selector is worth doing more work on. If the simple selectors don't match, or there's a child selector that does not appear in the bloom parent bloom filter, we can exit early.