Function selectors::matching::matches_compound_selector_from

pub fn matches_compound_selector_from<E>(
    selector: &Selector<E::Impl>, 
    from_offset: usize, 
    context: &mut MatchingContext<'_, E::Impl>, 
    element: &E
) -> CompoundSelectorMatchingResult where
    E: Element, 

Matches a compound selector belonging to selector, starting at offset from_offset, matching left to right.

Requires that from_offset points to a Combinator.

NOTE(emilio): This doesn’t allow to match in the leftmost sequence of the complex selector, but it happens to be the case we don’t need it.