Struct selectors::matching::MatchingContext

pub struct MatchingContext<'a> {
    pub relations: StyleRelations,
    pub matching_mode: MatchingMode,
    pub bloom_filter: Option<&'a BloomFilter>,
}

Data associated with the matching process for a element. This context is used across many selectors for an element, so it's not appropriate for transient data that applies to only a single selector.

Fields
relations: StyleRelations

Output that records certains relations between elements noticed during matching (and also extended after matching).

matching_mode: MatchingMode

The matching mode we should use when matching selectors.

bloom_filter: Option<&'a BloomFilter>

The bloom filter used to fast-reject selectors.

Methods

impl<'a> MatchingContext<'a>

pub fn new(
matching_mode: MatchingMode,
bloom_filter: Option<&'a BloomFilter>
) -> Self

Constructs a new MatchingContext.