Enum selectors::parser::Combinator

pub enum Combinator {
    Child,
    Descendant,
    NextSibling,
    LaterSibling,
    PseudoElement,
    SlotAssignment,
    Part,
}

Variants

Child

Descendant

NextSibling

LaterSibling

PseudoElement

A dummy combinator we use to the left of pseudo-elements.

It serializes as the empty string, and acts effectively as a child combinator in most cases. If we ever actually start using a child combinator for this, we will need to fix up the way hashes are computed for revalidation selectors.

SlotAssignment

Another combinator used for ::slotted(), which represent the jump from a node to its assigned slot.

Part

Another combinator used for ::part(), which represents the jump from the part to the containing shadow host.

Implementations

impl Combinator

pub fn is_ancestor(&self) -> bool

Returns true if this combinator is a child or descendant combinator.

pub fn is_pseudo_element(&self) -> bool

Returns true if this combinator is a pseudo-element combinator.

pub fn is_sibling(&self) -> bool

Returns true if this combinator is a next- or later-sibling combinator.