Enum selectors::parser::RelativeSelectorMatchHint

pub enum RelativeSelectorMatchHint {
    InSubtree,
    InChild,
    InNextSibling,
    InNextSiblingSubtree,
    InSibling,
    InSiblingSubtree,
}

Flag indicating where a given relative selector’s match would be contained.

Variants

InSubtree

Within this element’s subtree.

InChild

Within this element’s direct children.

InNextSibling

This element’s next sibling.

InNextSiblingSubtree

Within this element’s next sibling’s subtree.

InSibling

Within this element’s subsequent siblings.

InSiblingSubtree

Across this element’s subsequent siblings and their subtrees.

Implementations

impl RelativeSelectorMatchHint

pub fn new( relative_combinator: Combinator, has_child_or_descendants: bool, has_adjacent_or_next_siblings: bool, ) -> Self

Create a new relative selector match hint based on its composition.

pub fn is_descendant_direction(&self) -> bool

Is the match traversal direction towards the descendant of this element (As opposed to siblings)?

pub fn is_next_sibling(&self) -> bool

Is the match traversal terminated at the next sibling?

pub fn is_subtree(&self) -> bool

Does the match involve matching the subtree?