Struct selectors::parser::ComplexSelector

pub struct ComplexSelector<Impl: SelectorImpl>(_);

A ComplexSelectors stores a sequence of simple selectors and combinators. The iterator classes allow callers to iterate at either the raw sequence level or at the level of sequences of simple selectors separated by combinators. Most callers want the higher-level iterator.

We store selectors internally left-to-right (in parsing order), but the canonical iteration order is right-to-left (selector matching order). The iterators abstract over these details.

Methods

impl<Impl: SelectorImpl> ComplexSelector<Impl>

Important traits for SelectorIter<'a, Impl>
impl<'a, Impl: SelectorImpl> Iterator for SelectorIter<'a, Impl> type Item = &'a Component<Impl>;

pub fn iter(&self) -> SelectorIter<Impl>

Returns an iterator over the next sequence of simple selectors. When a combinator is reached, the iterator will return None, and next_sequence() may be called to continue to the next sequence.

pub fn iter_raw(&self) -> Rev<Iter<Component<Impl>>>

Returns an iterator over the entire sequence of simple selectors and combinators, from right to left.

pub fn iter_raw_rev(&self) -> Iter<Component<Impl>>

Returns an iterator over the entire sequence of simple selectors and combinators, from left to right.

Important traits for AncestorIter<'a, Impl>
impl<'a, Impl: SelectorImpl> Iterator for AncestorIter<'a, Impl> type Item = &'a Component<Impl>;

pub fn iter_ancestors(&self) -> AncestorIter<Impl>

Returns an iterator over ancestor simple selectors. All combinators and non-ancestor simple selectors will be skipped.

pub fn slice_from(&self, index: usize) -> Self

Returns a ComplexSelector identical to |self| but with the rightmost |index| entries removed.

pub fn slice_to(&self, index: usize) -> Self

Returns a ComplexSelector identical to |self| but with the leftmost |len() - index| entries removed.

pub fn from_vec(vec: Vec<Component<Impl>>) -> Self

Creates a ComplexSelector from a vec of Components. Used in tests.

impl<Impl: SelectorImpl> ComplexSelector<Impl>

pub fn parse<P>(parser: &P, input: &mut CssParser) -> Result<Self, ()> where
P: Parser<Impl = Impl>,

Parse a complex selector, without any pseudo-element.