Struct selectors::parser::Selector

pub struct Selector<Impl: SelectorImpl>(_);

A Selector stores a sequence of simple selectors and combinators. The iterator classes allow callers to iterate at either the raw sequence level or at the level of sequences of simple selectors separated by combinators. Most callers want the higher-level iterator.

We store compound selectors internally right-to-left (in matching order). Additionally, we invert the order of top-level compound selectors so that each one matches left-to-right. This is because matching namespace, local name, id, and class are all relatively cheap, whereas matching pseudo-classes might be expensive (depending on the pseudo-class). Since authors tend to put the pseudo-classes on the right, it’s faster to start matching on the left.

This reordering doesn’t change the semantics of selector matching, and we handle it in to_css to make it invisible to serialization.

Implementations

impl<Impl: SelectorImpl> Selector<Impl>

pub fn specificity(&self) -> u32

pub fn has_pseudo_element(&self) -> bool

pub fn is_slotted(&self) -> bool

pub fn pseudo_element(&self) -> Option<&Impl::PseudoElement>

pub fn is_universal(&self) -> bool

Whether this selector (pseudo-element part excluded) matches every element.

Used for “pre-computed” pseudo-elements in components/style/stylist.rs

pub fn iter(&self) -> SelectorIter<'_, Impl>

Returns an iterator over this selector in matching order (right-to-left). When a combinator is reached, the iterator will return None, and next_sequence() may be called to continue to the next sequence.

pub fn is_featureless_host_selector_or_pseudo_element(&self) -> bool

Whether this selector is a featureless :host selector, with no combinators to the left, and optionally has a pseudo-element to the right.

pub fn iter_from(&self, offset: usize) -> SelectorIter<'_, Impl>

Returns an iterator over this selector in matching order (right-to-left), skipping the rightmost |offset| Components.

pub fn combinator_at_match_order(&self, index: usize) -> Combinator

Returns the combinator at index index (zero-indexed from the right), or panics if the component is not a combinator.

pub fn iter_raw_match_order(&self) -> Iter<'_, Component<Impl>>

Returns an iterator over the entire sequence of simple selectors and combinators, in matching order (from right to left).

pub fn combinator_at_parse_order(&self, index: usize) -> Combinator

Returns the combinator at index index (zero-indexed from the left), or panics if the component is not a combinator.

pub fn iter_raw_parse_order_from(
&self,
offset: usize
) -> Rev<Iter<'_, Component<Impl>>>

Returns an iterator over the sequence of simple selectors and combinators, in parse order (from left to right), starting from offset.

pub fn from_vec(vec: Vec<Component<Impl>>, specificity_and_flags: u32) -> Self

Creates a Selector from a vec of Components, specified in parse order. Used in tests.

pub fn len(&self) -> usize

Returns count of simple selectors and combinators in the Selector.

pub fn thin_arc_heap_ptr(&self) -> *const c_void

Returns the address on the heap of the ThinArc for memory reporting.

impl<Impl: SelectorImpl> Selector<Impl>

pub fn parse<'i, 't, P>(
parser: &P,
input: &mut CssParser<'i, 't>
) -> Result<Self, ParseError<'i, P::Error>> where
P: Parser<'i, Impl = Impl>,

Parse a selector, without any pseudo-element.