Struct selectors::parser::SelectorInner

pub struct SelectorInner<Impl: SelectorImpl> {
    pub complex: ComplexSelector<Impl>,
    pub ancestor_hashes: [u32; 4],
}

The cores parts of a selector used for matching. This exists to make that information accessibly separately from the specificity and pseudo-element information that lives on |Selector| proper. We may want to refactor things and move that information elsewhere, at which point we could rename this to |Selector|.

Fields
complex: ComplexSelector<Impl>

The selector data.

ancestor_hashes: [u32; 4]

Ancestor hashes for the bloom filter. We precompute these and store them inline to optimize cache performance during selector matching. This matters a lot.

Methods

impl<Impl: SelectorImpl> SelectorInner<Impl>

pub fn new(c: ComplexSelector<Impl>) -> Self

pub fn from_vec(vec: Vec<Component<Impl>>) -> Self

Creates a SelectorInner from a Vec of Components. Used in tests.