Struct selectors::relative_selector::filter::RelativeSelectorFilterMap

pub struct RelativeSelectorFilterMap {
    map: FxHashMap<Key, Entry>,
}

Map of bloom filters for fast-rejecting relative selectors.

Fields
map: FxHashMap<Key, Entry>
Implementations

impl RelativeSelectorFilterMap

fn get_filter<E: Element>( &mut self, element: &E, kind: TraversalKind ) -> Option<&BloomFilter>

pub fn fast_reject<Impl: SelectorImpl, E: Element>( &mut self, element: &E, selector: &RelativeSelector<Impl>, quirks_mode: QuirksMode ) -> bool

Potentially reject the given selector for this element. This may seem redundant in presence of the cache, but the cache keys into the selector-element pair specifically, while this filter keys to the element and the traversal kind, so it is useful for handling multiple selectors that effectively end up looking at the same(-ish, for siblings) subtree.