selectors::sink

Trait Push

pub trait Push<T> {
    // Required method
    fn push(&mut self, value: T);
}

A trait to abstract over a push method that may be implemented for different kind of types.

Used to abstract over Array, SmallVec and Vec, and also to implement a type which push method does only tweak a byte when we only need to check for the presence of something.

Required Methods

fn push(&mut self, value: T)

Push a value into self.