Module selectors::smallvec

Small vectors in various sizes. These store a certain number of elements inline and fall back to the heap for larger allocations.

Structs

SmallVec1

SmallVec16

SmallVec2

SmallVec24

SmallVec32

SmallVec4

SmallVec8

SmallVecIterator

SmallVecMoveIterator

SmallVecMutIterator

Traits

SmallVec

SmallVecPrivate

VecLike