Crate selectors

Re-exports

pub use crate::parser::Parser;

pub use crate::parser::SelectorImpl;

pub use crate::parser::SelectorList;

Modules

attr

bloom
Counting and non-counting Bloom filters tuned for use as ancestor filters for selector matching.

builder
Helper module to build up a selector safely and efficiently.

context

matching

nth_index_cache

parser

sink
Small helpers to abstract over different containers.

tree
Traits that nodes must implement. Breaks the otherwise-cyclic dependency between layout and style.

visitor
Visitor traits for selectors.

Structs

NthIndexCache
A cache to speed up matching of nth-index-like selectors.

OpaqueElement
Opaque representation of an Element, for identity comparisons.

Traits

Element