Module selectors::bloom

Simple counting bloom filters.

Structs

BloomFilter
A counting Bloom filter with 8-bit counters. For now we assume that having two hash functions is enough, but we may revisit that decision later.

Traits

BloomHash