Struct selectors::builder::SelectorBuilder

pub struct SelectorBuilder<Impl: SelectorImpl> {
    simple_selectors: SmallVec<[Component<Impl>; 32]>,
    combinators: SmallVec<[(Combinator, usize); 16]>,
    current_len: usize,
}

Top-level SelectorBuilder struct. This should be stack-allocated by the consumer and never moved (because it contains a lot of inline data that would be slow to memmov).

After instantation, callers may call the push_simple_selector() and push_combinator() methods to append selector data as it is encountered (from left to right). Once the process is complete, callers should invoke build(), which transforms the contents of the SelectorBuilder into a heap- allocated Selector and leaves the builder in a drained state.

Fields
simple_selectors: SmallVec<[Component<Impl>; 32]>

The entire sequence of simple selectors, from left to right, without combinators.

We make this large because the result of parsing a selector is fed into a new Arc-ed allocation, so any spilled vec would be a wasted allocation. Also, Components are large enough that we don’t have much cache locality benefit from reserving stack space for fewer of them.

combinators: SmallVec<[(Combinator, usize); 16]>

The combinators, and the length of the compound selector to their left.

current_len: usize

The length of the current compount selector.

Implementations

impl<Impl: SelectorImpl> SelectorBuilder<Impl>

pub fn push_simple_selector(&mut self, ss: Component<Impl>)

Pushes a simple selector onto the current compound selector.

pub fn push_combinator(&mut self, c: Combinator)

Completes the current compound selector and starts a new one, delimited by the given combinator.

pub fn has_combinators(&self) -> bool

Returns true if combinators have ever been pushed to this builder.

pub fn build(
&mut self,
parsed_pseudo: bool,
parsed_slotted: bool,
parsed_part: bool
) -> ThinArc<SpecificityAndFlags, Component<Impl>>

Consumes the builder, producing a Selector.

pub fn build_with_specificity_and_flags(
&mut self,
spec: SpecificityAndFlags
) -> ThinArc<SpecificityAndFlags, Component<Impl>>

Builds with an explicit SpecificityAndFlags. This is separated from build() so that unit tests can pass an explicit specificity.