Module selectors::context

Structs

MatchingContext

Data associated with the matching process for a element. This context is used across many selectors for an element, so it’s not appropriate for transient data that applies to only a single selector.

Enums

MatchingMode

What kind of selector matching mode we should use.

QuirksMode

Which quirks mode is this document in.

VisitedHandlingMode

The mode to use when matching unvisited and visited links.