selectors

Module matching

Re-exports

pub use crate::context::*;

Structs

ElementSelectorFlags

Set of flags that are set on either the element or its parent (depending on the flag) if the element could potentially match a selector.

LocalMatchingContext 🔒

Holds per-compound-selector data.

NextElement 🔒

Enums

CompoundSelectorMatchingResult

Whether a compound selector matched, and whether it was the rightmost selector inside the complex selector.

SelectorMatchingResult 🔒

A result of selector matching, includes 3 failure types,

SubjectOrPseudoElement 🔒

Statics

RECOMMENDED_SELECTOR_BLOOM_FILTER_SIZE

Functions

assigned_slot 🔒

compound_matches_featureless_host 🔒

There are relatively few selectors in a given compound that may match a featureless element. Instead of adding a check to every selector that may not match, we handle it here in an out of line path.

do_match_relative_selectors 🔒

Matches a relative selector in a list of relative selectors.

host_for_part 🔒

hover_and_active_quirk_applies 🔒

Whether the :hover and :active quirk applies.

match_relative_selectors 🔒

matches_complex_selector 🔒

Matches a complex selector.

matches_complex_selector_internal 🔒

matches_complex_selector_list 🔒

Matches each selector of a list as a complex selector

matches_compound_selector 🔒

Determines whether the given element matches the given compound selector.

matches_compound_selector_from

Matches a compound selector belonging to selector, starting at offset from_offset, matching left to right.

matches_generic_nth_child 🔒

matches_host 🔒

matches_local_name 🔒

matches_part 🔒

matches_rare_attribute_selector 🔒

matches_relative_selector 🔒

matches_relative_selector_subtree 🔒

matches_selector

Matches a selector, fast-rejecting against a bloom filter.

matches_selector_kleene

Same as matches_selector, but returns the Kleene value as-is.

matches_selector_list

matches_simple_selector 🔒

Determines whether the given element matches the given single selector.

matches_slotted 🔒

next_element_for_combinator 🔒

nth_child_index 🔒

relative_selector_match_early 🔒

select_name

selector_may_match

Given the ancestor hashes from a selector, see if the current element, represented by the bloom filter, has a chance of matching at all.

to_unconditional_case_sensitivity