Module selectors::matching

Structs

CommonStyleAffectingAttributeInfo

CommonStyleAffectingAttributes

DeclarationBlock
A property declaration together with its precedence among rules of equal specificity so that we can sort them.

ElementFlags
Flags set on elements during the matching process.

Rule

SelectorMap
Map element data to Rules whose last simple selector starts with them.

Enums

CommonStyleAffectingAttributeMode

Constants

ALIGN_CENTER_ATTRIBUTE

ALIGN_LEFT_ATTRIBUTE

ALIGN_RIGHT_ATTRIBUTE

HAS_EDGE_CHILD_SELECTOR
When a child is added or removed from this element, the first and last children must be restyled, because they may match :first-child, :last-child, or :only-child.

HAS_SLOW_SELECTOR
When a child is added or removed from this element, all the children must be restyled, because they may match :nth-last-child, :last-of-type, :nth-last-of-type, or :only-of-type.

HAS_SLOW_SELECTOR_LATER_SIBLINGS
When a child is added or removed from this element, any later children must be restyled, because they may match :nth-child, :first-of-type, or :nth-of-type.

HIDDEN_ATTRIBUTE

NO_WRAP_ATTRIBUTE

Statics

RECOMMENDED_SELECTOR_BLOOM_FILTER_SIZE

SELECTOR_WHITESPACE
The definition of whitespace per CSS Selectors Level 3  4.

Functions

common_style_affecting_attributes

matches

matches_compound_selector
Determines whether the given element matches the given single or compound selector.

matches_simple_selector
Determines whether the given element matches the given single selector.

rare_style_affecting_attributes
Attributes that, if present, disable style sharing. All legacy HTML attributes must be in either this list or common_style_affecting_attributes. See the comment in synthesize_presentational_hints_for_legacy_attributes.