Module selectors::matching

Re-exports

pub use crate::context::*;

Structs

ElementSelectorFlags
Set of flags that are set on either the element or its parent (depending on the flag) if the element could potentially match a selector.

LocalMatchingContext
Holds per-compound-selector data.

Enums

CompoundSelectorMatchingResult
Whether a compound selector matched, and whether it was the rightmost selector inside the complex selector.

MatchesHoverAndActiveQuirk
Whether the :hover and :active quirk applies.

Rightmost

SelectorMatchingResult
A result of selector matching, includes 3 failure types,

Statics

RECOMMENDED_SELECTOR_BLOOM_FILTER_SIZE

Functions

matches_complex_selector
Matches a complex selector.

matches_complex_selector_internal

matches_compound_selector
Determines whether the given element matches the given compound selector.

matches_compound_selector_from
Matches a compound selector belonging to selector, starting at offset from_offset, matching left to right.

matches_first_child

matches_generic_nth_child

matches_hover_and_active_quirk

matches_last_child

matches_local_name

matches_selector
Matches a selector, fast-rejecting against a bloom filter.

matches_selector_list

matches_simple_selector
Determines whether the given element matches the given single selector.

may_match

next_element_for_combinator

nth_child_index

select_name