Module selectors::matching

Structs

CommonStyleAffectingAttributeInfo

CommonStyleAffectingAttributes

DeclarationBlock
A property declaration together with its precedence among rules of equal specificity so that we can sort them.

Rule

SelectorMap
Map node attributes to Rules whose last simple selector starts with them.

Enums

CommonStyleAffectingAttributeMode

Constants

ALIGN_CENTER_ATTRIBUTE

ALIGN_LEFT_ATTRIBUTE

ALIGN_RIGHT_ATTRIBUTE

HIDDEN_ATTRIBUTE

NO_WRAP_ATTRIBUTE

Statics

RECOMMENDED_SELECTOR_BLOOM_FILTER_SIZE

SELECTOR_WHITESPACE
The definition of whitespace per CSS Selectors Level 3  4.

Functions

common_style_affecting_attributes

matches

matches_simple_selector
Determines whether the given element matches the given single selector.

rare_style_affecting_attributes
Attributes that, if present, disable style sharing. All legacy HTML attributes must be in either this list or common_style_affecting_attributes. See the comment in synthesize_presentational_hints_for_legacy_attributes.