Module selectors::matching

Reexports

pub use context::*;

Structs

ElementSelectorFlags
Set of flags that are set on either the element or its parent (depending on the flag) if the element could potentially match a selector.

Enums

CompoundSelectorMatchingResult
Whether a compound selector matched, and whether it was the rightmost selector inside the complex selector.

Statics

RECOMMENDED_SELECTOR_BLOOM_FILTER_SIZE

Functions

matches_complex_selector
Matches a complex selector.

matches_compound_selector
Matches a compound selector belonging to selector, starting at offset from_offset, matching left to right.

matches_selector
Matches a selector, fast-rejecting against a bloom filter.

matches_selector_list