Struct selectors::matching::SelectorMap

pub struct SelectorMap<T, Impl: SelectorImpl> {
    id_hash: HashMap<Atom, Vec<Rule<T, Impl>>>,
    class_hash: HashMap<Atom, Vec<Rule<T, Impl>>>,
    local_name_hash: HashMap<Atom, Vec<Rule<T, Impl>>>,
    lower_local_name_hash: HashMap<Atom, Vec<Rule<T, Impl>>>,
    other_rules: Vec<Rule<T, Impl>>,
    empty: bool,
}

Map element data to Rules whose last simple selector starts with them.

e.g., "p > img" would go into the set of Rules corresponding to the element "img" "a .foo .bar.baz" would go into the set of Rules corresponding to the class "bar"

Because we match Rules right-to-left (i.e., moving up the tree from an element), we need to compare the last simple selector in the Rule with the element.

So, if an element has ID "id1" and classes "foo" and "bar", then all the rules it matches will have their last simple selector starting either with "#id1" or with ".foo" or with ".bar".

Hence, the union of the rules keyed on each of element's classes, ID, element name, etc. will contain the Rules that actually match that element.

Fields

id_hash

class_hash

local_name_hash

lower_local_name_hash

Same as local_name_hash, but keys are lower-cased. For HTML elements in HTML documents.

other_rules

Rules that don't have ID, class, or element selectors.

empty

Whether this hash is empty.

Methods

impl<T, Impl: SelectorImpl> SelectorMap<T, Impl>

fn new() -> SelectorMap<T, Impl>

fn get_all_matching_rules<E, V>(&self, element: &E, parent_bf: Option<&BloomFilter>, matching_rules_list: &mut V, shareable: &mut bool) where E: Element<Impl=Impl>, V: VecLike<DeclarationBlock<T>>

Append to rule_list all Rules in self that match element.

Extract matching rules as per element's ID, classes, tag name, etc.. Sort the Rules at the end to maintain cascading order.

fn get_universal_rules<V>(&self, matching_rules_list: &mut V) where V: VecLike<DeclarationBlock<T>>

Append to rule_list all universal Rules (rules with selector *|*) in self sorted by specifity and source order.

fn get_matching_rules_from_hash<E, V>(element: &E, parent_bf: Option<&BloomFilter>, hash: &HashMap<Atom, Vec<Rule<T, Impl>>>, key: &Atom, matching_rules: &mut V, shareable: &mut bool) where E: Element<Impl=Impl>, V: VecLike<DeclarationBlock<T>>

fn get_matching_rules<E, V>(element: &E, parent_bf: Option<&BloomFilter>, rules: &[Rule<T, Impl>], matching_rules: &mut V, shareable: &mut bool) where E: Element<Impl=Impl>, V: VecLike<DeclarationBlock<T>>

Adds rules in rules that match element to the matching_rules list.

fn insert(&mut self, rule: Rule<T, Impl>)

Insert rule into the correct hash. Order in which to try: id_hash, class_hash, local_name_hash, other_rules.

fn get_id_name(rule: &Rule<T, Impl>) -> Option<Atom>

Retrieve the first ID name in Rule, or None otherwise.

fn get_class_name(rule: &Rule<T, Impl>) -> Option<Atom>

Retrieve the FIRST class name in Rule, or None otherwise.

fn get_local_name(rule: &Rule<T, Impl>) -> Option<LocalName>

Retrieve the name if it is a type selector, or None otherwise.