Struct selectors::matching::SelectorMap

pub struct SelectorMap<T> {
    // some fields omitted
}

Map node attributes to Rules whose last simple selector starts with them.

e.g., "p > img" would go into the set of Rules corresponding to the element "img" "a .foo .bar.baz" would go into the set of Rules corresponding to the class "bar"

Because we match Rules right-to-left (i.e., moving up the tree from a node), we need to compare the last simple selector in the Rule with the node.

So, if a node has ID "id1" and classes "foo" and "bar", then all the rules it matches will have their last simple selector starting either with "#id1" or with ".foo" or with ".bar".

Hence, the union of the rules keyed on each of node's classes, ID, element name, etc. will contain the Rules that actually match that node.

Methods

impl<T> SelectorMap<T>

fn new() -> SelectorMap<T>

fn get_all_matching_rules<'a, N, V>(&self, node: &N, parent_bf: &Option<Box<BloomFilter>>, matching_rules_list: &mut V, shareable: &mut bool) where N: TNode<'a>, V: VecLike<DeclarationBlock<T>>

Append to rule_list all Rules in self that match node.

Extract matching rules as per node's ID, classes, tag name, etc.. Sort the Rules at the end to maintain cascading order.

fn insert(&mut self, rule: Rule<T>)

Insert rule into the correct hash. Order in which to try: id_hash, class_hash, local_name_hash, universal_rules.