Module selectors::parser

Re-exports

pub use crate::visitor::SelectorVisitor;

pub use crate::visitor::Visit;

Structs

AncestorHashes

Ancestor hashes for the bloom filter. We precompute these and store them inline with selectors to optimize cache performance during matching. This matters a lot.

LocalName

Selector

A Selector stores a sequence of simple selectors and combinators. The iterator classes allow callers to iterate at either the raw sequence level or at the level of sequences of simple selectors separated by combinators. Most callers want the higher-level iterator.

SelectorIter

SelectorList

Enums

Combinator

Component

A CSS simple selector or combinator. We store both in the same enum for optimal packing and cache performance, see [1].

SelectorParseErrorKind

Traits

NonTSPseudoClass

A trait that represents a pseudo-class.

Parser

PseudoElement

A trait that represents a pseudo-element.

SelectorImpl

This trait allows to define the parser implementation in regards of pseudo-classes/elements

Functions

namespace_empty_string

parse_compound_selector_list

Parse a comma separated list of compound selectors.

Type Aliases

SelectorParseError