Enum selectors::parser::Component

pub enum Component<Impl: SelectorImpl> {

    Combinator(Combinator),
    ExplicitAnyNamespace,
    ExplicitNoNamespace,
    DefaultNamespace(Impl::NamespaceUrl),
    Namespace(Impl::NamespacePrefix, Impl::NamespaceUrl),
    ExplicitUniversalType,
    LocalName(LocalName<Impl>),
    ID(Impl::Identifier),
    Class(Impl::Identifier),
    AttributeInNoNamespaceExists {
        local_name: Impl::LocalName,
        local_name_lower: Impl::LocalName,
    },
    AttributeInNoNamespace {
        local_name: Impl::LocalName,
        operator: AttrSelectorOperator,
        value: Impl::AttrValue,
        case_sensitivity: ParsedCaseSensitivity,
        never_matches: bool,
    },
    AttributeOther(Box<AttrSelectorWithOptionalNamespace<Impl>>),
    Negation(Box<[Selector<Impl>]>),
    FirstChild,
    LastChild,
    OnlyChild,
    Root,
    Empty,
    Scope,
    NthChild(i32, i32),
    NthLastChild(i32, i32),
    NthOfType(i32, i32),
    NthLastOfType(i32, i32),
    FirstOfType,
    LastOfType,
    OnlyOfType,
    NonTSPseudoClass(Impl::NonTSPseudoClass),
    Slotted(Selector<Impl>),
    Part(Box<[Impl::Identifier]>),
    Host(Option<Selector<Impl>>),
    Where(Box<[Selector<Impl>]>),
    Is(Box<[Selector<Impl>]>),
    PseudoElement(Impl::PseudoElement),

}

A CSS simple selector or combinator. We store both in the same enum for optimal packing and cache performance, see [1].

[1] https://bugzilla.mozilla.org/show_bug.cgi?id=1357973

Variants

Combinator

Tuple Fields of Combinator

0: Combinator

ExplicitAnyNamespace

ExplicitNoNamespace

DefaultNamespace

Tuple Fields of DefaultNamespace

0: Impl::NamespaceUrl

Namespace

Tuple Fields of Namespace

0: Impl::NamespacePrefix1: Impl::NamespaceUrl

ExplicitUniversalType

LocalName

Tuple Fields of LocalName

0: LocalName<Impl>

ID

Tuple Fields of ID

0: Impl::Identifier

Class

Tuple Fields of Class

0: Impl::Identifier

AttributeInNoNamespaceExists

Fields of AttributeInNoNamespaceExists

local_name: Impl::LocalNamelocal_name_lower: Impl::LocalName

AttributeInNoNamespace

Fields of AttributeInNoNamespace

local_name: Impl::LocalNameoperator: AttrSelectorOperatorvalue: Impl::AttrValuecase_sensitivity: ParsedCaseSensitivitynever_matches: bool

AttributeOther

Tuple Fields of AttributeOther

0: Box<AttrSelectorWithOptionalNamespace<Impl>>

Negation

Pseudo-classes

Tuple Fields of Negation

0: Box<[Selector<Impl>]>

FirstChild

LastChild

OnlyChild

Root

Empty

Scope

NthChild

Tuple Fields of NthChild

0: i321: i32

NthLastChild

Tuple Fields of NthLastChild

0: i321: i32

NthOfType

Tuple Fields of NthOfType

0: i321: i32

NthLastOfType

Tuple Fields of NthLastOfType

0: i321: i32

FirstOfType

LastOfType

OnlyOfType

NonTSPseudoClass

Tuple Fields of NonTSPseudoClass

0: Impl::NonTSPseudoClass

Slotted

The ::slotted() pseudo-element:

https://drafts.csswg.org/css-scoping/#slotted-pseudo

The selector here is a compound selector, that is, no combinators.

NOTE(emilio): This should support a list of selectors, but as of this writing no other browser does, and that allows them to put ::slotted() in the rule hash, so we do that too.

See https://github.com/w3c/csswg-drafts/issues/2158

Tuple Fields of Slotted

0: Selector<Impl>

Part

The ::part pseudo-element. https://drafts.csswg.org/css-shadow-parts/#part

Tuple Fields of Part

0: Box<[Impl::Identifier]>

Host

The :host pseudo-class:

https://drafts.csswg.org/css-scoping/#host-selector

NOTE(emilio): This should support a list of selectors, but as of this writing no other browser does, and that allows them to put :host() in the rule hash, so we do that too.

See https://github.com/w3c/csswg-drafts/issues/2158

Tuple Fields of Host

0: Option<Selector<Impl>>

Where

The :where pseudo-class.

https://drafts.csswg.org/selectors/#zero-matches

The inner argument is conceptually a SelectorList, but we move the selectors to the heap to keep Component small.

Tuple Fields of Where

0: Box<[Selector<Impl>]>

Is

The :is pseudo-class.

https://drafts.csswg.org/selectors/#matches-pseudo

Same comment as above re. the argument.

Tuple Fields of Is

0: Box<[Selector<Impl>]>

PseudoElement

An implementation-dependent pseudo-element selector.

Tuple Fields of PseudoElement

0: Impl::PseudoElement

Implementations

impl<Impl: SelectorImpl> Component<Impl>

pub fn is_combinator(&self) -> bool

Returns true if this is a combinator.

pub fn as_combinator(&self) -> Option<Combinator>

Returns the value as a combinator if applicable, None otherwise.

pub fn maybe_allowed_after_pseudo_element(&self) -> bool

Whether this component is valid after a pseudo-element. Only intended for sanity-checking.

fn matches_for_stateless_pseudo_element(&self) -> bool

Whether a given selector should match for stateless pseudo-elements.

This is a bit subtle: Only selectors that return true in maybe_allowed_after_pseudo_element should end up here, and NonTSPseudoClass never matches (as it is a stateless pseudo after all).

pub fn visit<V>(&self, visitor: &mut V) -> bool where
V: SelectorVisitor<Impl = Impl>,