Trait selectors::visitor::SelectorVisitor

pub trait SelectorVisitor: Sized {
    type Impl: SelectorImpl;
    fn visit_attribute_selector(
        &mut self, 
        _namespace: &NamespaceConstraint<&<Self::Impl as SelectorImpl>::NamespaceUrl>, 
        _local_name: &<Self::Impl as SelectorImpl>::LocalName, 
        _local_name_lower: &<Self::Impl as SelectorImpl>::LocalName
    ) -> bool { ... }

    fn visit_simple_selector(&mut self, _: &Component<Self::Impl>) -> bool { ... }

    fn visit_selector_list(&mut self, list: &[Selector<Self::Impl>]) -> bool { ... }

    fn visit_complex_selector(
        &mut self, 
        _combinator_to_right: Option<Combinator>
    ) -> bool { ... }
}

A trait to visit selector properties.

All the visit_foo methods return a boolean indicating whether the traversal should continue or not.

Associated Types

type Impl: SelectorImpl

The selector implementation this visitor wants to visit.

Provided methods

fn visit_attribute_selector(
&mut self,
_namespace: &NamespaceConstraint<&<Self::Impl as SelectorImpl>::NamespaceUrl>,
_local_name: &<Self::Impl as SelectorImpl>::LocalName,
_local_name_lower: &<Self::Impl as SelectorImpl>::LocalName
) -> bool

Visit an attribute selector that may match (there are other selectors that may never match, like those containing whitespace or the empty string).

fn visit_simple_selector(&mut self, _: &Component<Self::Impl>) -> bool

Visit a simple selector.

fn visit_selector_list(&mut self, list: &[Selector<Self::Impl>]) -> bool

Visit a nested selector list. The caller is responsible to call visit into the internal selectors if / as needed.

The default implementation does this.

fn visit_complex_selector(
&mut self,
_combinator_to_right: Option<Combinator>
) -> bool

Visits a complex selector.

Gets the combinator to the right of the selector, or None if the selector is the rightmost one.